Extract Checkbox and RadioGroup
