+++ src/atomic/atoms/RadioGroup.jsx
@@ -3,30 +3,13 @@ import React from 'react';
 import {
   FormControl,
   // FormHelperText,
-  FormControlLabel,
+  // FormControlLabel,
 } from 'material-ui/Form';
 
 // import { InputLabel as MUIInputLabel } from 'material-ui/Input';
-import MUITextField from 'material-ui/TextField';
+// import MUITextField from 'material-ui/TextField';
 import { RadioGroup as MUIRadioGroup } from 'material-ui/Radio';
-import MUICheckbox from 'material-ui/Checkbox';
-
-const noop = () => {};
-
-
-
-export const Checkbox = ({ input, meta: { touched, error }, label }) => (
-  <FormControlLabel
-    error={touched && error ? error : null}
-    control={
-      <MUICheckbox
-        checked={input.value ? true : false}
-        onChange={input.onChange}
-      />
-    }
-    label={label}
-  />
-);
+// import MUICheckbox from 'material-ui/Checkbox';
 
 export const RadioGroup = ({ input, meta: { touched, error }, ...rest }) => (
   <FormControl error={Boolean(touched && error)}>
@@ -38,3 +21,5 @@ export const RadioGroup = ({ input, meta: { touched, error }, ...rest }) => (
     />
   </FormControl>
 );
+
+export default RadioGroup;
