+++ src/atomic/index.js
@@ -1,3 +1,4 @@
-export * from './atoms/MUI-Wrappers';
 export { default as SelectField } from './atoms/SelectField';
 export { default as TextField } from './atoms/TextField';
+export { default as Checkbox } from './atoms/Checkbox';
+export { default as RadioGroup } from './atoms/RadioGroup';
